added minimal example of registering a pointer with a resource manager
This example is a use case from one of our internal programs at PNNL. We often recieve
data (both on the host and device) from a third party library, and would then like
to move that data to/from the host and device. Registering the externally allocated
data as demonstrated in this example is how we acomplished this.